Move RiskList into components
